Create Generatable module and add methods
